Create package record
