+++ request/collect.go
@@ -1,7 +1,9 @@
 package request
 
-func collect(c <-chan Record) []Record {
-	rec := []Record{}
+import "github.com/benchttp/runner/record"
+
+func collect(c <-chan record.Record) []record.Record {
+	rec := []record.Record{}
 
 	for r := range c {
 		rec = append(rec, r)
